+++ json_serializable/lib/src/type_helpers/json_converter_helper.dart
@@ -143,6 +143,24 @@ class _JsonConvertData {
       accessor.isEmpty ? '' : '.$accessor';
 }
 
+/// If there is no converter for the params, return `null`.
+///
+/// Otherwise, returns `true` if the converter has a null return value.
+///
+/// Used to make sure we create a smart encoding function.
+bool? hasConverterNullEncode(
+  DartType targetType,
+  TypeHelperContextWithConfig ctx,
+) {
+  final data = _typeConverter(targetType, ctx);
+
+  if (data == null) {
+    return null;
+  }
+
+  return data.jsonType.isNullableType;
+}
+
 _JsonConvertData? _typeConverter(
   DartType targetType,
   TypeHelperContextWithConfig ctx,
@@ -199,7 +217,7 @@ _JsonConvertData? _typeConverterFrom(
 
   final annotationElement = match.elementAnnotation?.element;
   if (annotationElement is PropertyAccessorElement) {
-    final enclosing = annotationElement.enclosingElement3;
+    final enclosing = annotationElement.enclosingElement;
 
     var accessString = annotationElement.name;
 
@@ -226,7 +244,7 @@ _JsonConvertData? _typeConverterFrom(
 
   if (match.genericTypeArg != null) {
     return _JsonConvertData.genericClass(
-      match.annotation.type!.element2!.name!,
+      match.annotation.type!.element!.name!,
       match.genericTypeArg!,
       reviver.accessor,
       match.jsonType,
@@ -235,7 +253,7 @@ _JsonConvertData? _typeConverterFrom(
   }
 
   return _JsonConvertData.className(
-    match.annotation.type!.element2!.name!,
+    match.annotation.type!.element!.name!,
     reviver.accessor,
     match.jsonType,
     match.fieldType,
@@ -263,18 +281,18 @@ _ConverterMatch? _compatibleMatch(
   ElementAnnotation? annotation,
   DartObject constantValue,
 ) {
-  final converterClassElement = constantValue.type!.element2 as ClassElement;
+  final converterClassElement = constantValue.type!.element as ClassElement;
 
   final jsonConverterSuper =
       converterClassElement.allSupertypes.singleWhereOrNull(
-    (e) => _jsonConverterChecker.isExactly(e.element2),
+    (e) => _jsonConverterChecker.isExactly(e.element),
   );
 
   if (jsonConverterSuper == null) {
     return null;
   }
 
-  assert(jsonConverterSuper.element2.typeParameters.length == 2);
+  assert(jsonConverterSuper.element.typeParameters.length == 2);
   assert(jsonConverterSuper.typeArguments.length == 2);
 
   final fieldType = jsonConverterSuper.typeArguments[0];
@@ -305,7 +323,7 @@ _ConverterMatch? _compatibleMatch(
       annotation,
       constantValue,
       jsonConverterSuper.typeArguments[1],
-      '${targetType.element2.name}${targetType.isNullableType ? '?' : ''}',
+      '${targetType.element.name}${targetType.isNullableType ? '?' : ''}',
       fieldType,
     );
   }