+++ json_serializable/lib/src/utils.dart
@@ -160,16 +160,15 @@ ConstructorElement constructorByName(ClassElement classElement, String name) {
 ///
 /// Otherwise, `null`.
 Iterable<FieldElement>? iterateEnumFields(DartType targetType) {
-  if (targetType is InterfaceType && targetType.element2 is EnumElement) {
-    return targetType.element2.fields
-        .where((element) => element.isEnumConstant);
+  if (targetType is InterfaceType && targetType.element is EnumElement) {
+    return targetType.element.fields.where((element) => element.isEnumConstant);
   }
   return null;
 }
 
 extension DartTypeExtension on DartType {
   DartType promoteNonNullable() =>
-      element2?.library?.typeSystem.promoteToNonNull(this) ?? this;
+      element?.library?.typeSystem.promoteToNonNull(this) ?? this;
 }
 
 String ifNullOrElse(String test, String ifNull, String ifNotNull) =>
@@ -207,7 +206,7 @@ String typeToCode(
     return 'dynamic';
   } else if (type is InterfaceType) {
     return [
-      type.element2.name,
+      type.element.name,
       if (type.typeArguments.isNotEmpty)
         '<${type.typeArguments.map(typeToCode).join(', ')}>',
       (type.isNullableType || forceNullable) ? '?' : '',
@@ -229,7 +228,15 @@ extension ExecutableElementExtension on ExecutableElement {
     }
 
     if (this is MethodElement) {
-      return '${enclosingElement3.name}.$name';
+      return '${enclosingElement.name}.$name';
+    }
+
+    if (this is ConstructorElement) {
+      // Ignore the default constructor.
+      if (name.isEmpty) {
+        return '${enclosingElement.name}';
+      }
+      return '${enclosingElement.name}.$name';
     }
 
     throw UnsupportedError(