+++ json_serializable/lib/type_helper.dart
@@ -14,7 +14,7 @@ export 'src/type_helpers/enum_helper.dart';
 export 'src/type_helpers/iterable_helper.dart';
 export 'src/type_helpers/json_converter_helper.dart';
 export 'src/type_helpers/json_helper.dart';
-export 'src/type_helpers/map_helper.dart';
+export 'src/type_helpers/map_helper.dart' show MapHelper;
 export 'src/type_helpers/uri_helper.dart';
 export 'src/type_helpers/value_helper.dart';
 export 'src/unsupported_type_error.dart';