+++ json_serializable/test/default_value/implicit_default_value.dart
@@ -46,6 +46,12 @@ class DefaultValueImplicit implements dvi.DefaultValue {
   @JsonKey(fromJson: constClassFromJson, toJson: constClassToJson)
   ConstClass valueFromFunction;
 
+  int intDefaultValueFromFunction;
+
+  ConstClass valueFromDefaultValueDefaultConstructor;
+
+  ConstClass valueFromDefaultValueNamedConstructor;
+
   DefaultValueImplicit({
     this.fieldBool = true,
     this.fieldString = 'string',
@@ -65,6 +71,9 @@ class DefaultValueImplicit implements dvi.DefaultValue {
     this.constClass = const ConstClass('value'),
     this.valueFromConverter = const ConstClass('value'),
     this.valueFromFunction = const ConstClass('value'),
+    this.intDefaultValueFromFunction = 43,
+    this.valueFromDefaultValueDefaultConstructor = const ConstClass(),
+    this.valueFromDefaultValueNamedConstructor = const ConstClass('easy'),
   });
 
   factory DefaultValueImplicit.fromJson(Map<String, dynamic> json) =>