+++ json_serializable/test/default_value/implicit_default_value.g.dart
@@ -54,6 +54,20 @@ DefaultValueImplicit _$DefaultValueImplicitFromJson(
       valueFromFunction: json['valueFromFunction'] == null
           ? const ConstClass('value')
           : constClassFromJson(json['valueFromFunction'] as String),
+      intDefaultValueFromFunction:
+          json['intDefaultValueFromFunction'] as int? ?? 43,
+      valueFromDefaultValueDefaultConstructor:
+          json['valueFromDefaultValueDefaultConstructor'] == null
+              ? const ConstClass()
+              : ConstClass.fromJson(
+                  json['valueFromDefaultValueDefaultConstructor']
+                      as Map<String, dynamic>),
+      valueFromDefaultValueNamedConstructor:
+          json['valueFromDefaultValueNamedConstructor'] == null
+              ? const ConstClass('easy')
+              : ConstClass.fromJson(
+                  json['valueFromDefaultValueNamedConstructor']
+                      as Map<String, dynamic>),
     );
 
 Map<String, dynamic> _$DefaultValueImplicitToJson(
@@ -76,6 +90,11 @@ Map<String, dynamic> _$DefaultValueImplicitToJson(
       'valueFromConverter':
           const ConstClassConverter().toJson(instance.valueFromConverter),
       'valueFromFunction': constClassToJson(instance.valueFromFunction),
+      'intDefaultValueFromFunction': instance.intDefaultValueFromFunction,
+      'valueFromDefaultValueDefaultConstructor':
+          instance.valueFromDefaultValueDefaultConstructor,
+      'valueFromDefaultValueNamedConstructor':
+          instance.valueFromDefaultValueNamedConstructor,
     };
 
 const _$GreekEnumMap = {