+++ json_serializable/tool/test_type_builder.dart
@@ -7,6 +7,7 @@ import 'dart:async';
 import 'package:build/build.dart';
 import 'package:collection/collection.dart';
 import 'package:dart_style/dart_style.dart';
+import 'package:json_serializable/src/type_helpers/map_helper.dart';
 
 import 'shared.dart';
 import 'test_type_data.dart';
@@ -14,7 +15,7 @@ import 'test_type_data.dart';
 final _formatter = DartFormatter();
 
 const _trivialTypesToTest = {
-  'BigInt': TestTypeData(
+  'BigInt': TestTypeData.defaultFunc(
     jsonExpression: "'12345'",
     altJsonExpression: "'67890'",
   ),
@@ -22,7 +23,7 @@ const _trivialTypesToTest = {
     defaultExpression: 'true',
     altJsonExpression: 'false',
   ),
-  'DateTime': TestTypeData(
+  'DateTime': TestTypeData.defaultFunc(
     jsonExpression: "'2020-01-01T00:00:00.000'",
     altJsonExpression: "'2018-01-01T00:00:00.000'",
   ),
@@ -55,7 +56,7 @@ const _trivialTypesToTest = {
     defaultExpression: "'a string'",
     altJsonExpression: "'another string'",
   ),
-  'Uri': TestTypeData(
+  'Uri': TestTypeData.defaultFunc(
     jsonExpression: "'https://example.com'",
     altJsonExpression: "'https://dart.dev'",
   ),
@@ -96,16 +97,9 @@ final _typesToTest = {
   ..._collectionTypes,
 };
 
-const mapKeyTypes = {
-  'BigInt',
-  'DateTime',
-  'dynamic',
-  customEnumType,
-  'int',
-  'Object',
-  'String',
-  'Uri',
-};
+Iterable<String> get mapKeyTypes =>
+    allowedMapKeyTypes.map((e) => e == 'enum' ? customEnumType : e).toList()
+      ..sort(compareAsciiLowerCase);
 
 final _iterableGenericArgs = ([
   ..._trivialTypesToTest.keys,