+++ json_serializable/CHANGELOG.md
@@ -1,3 +1,40 @@
+## 6.6.0-dev
+
+- Require Dart SDK `>=2.18.0`.
+- Require `analyzer: ^5.2.0`
+
+## 6.5.4
+
+- Fixed handling of nullable fields with converters which return non-nullable
+  values.
+
+## 6.5.3
+
+- Fixed handling of nullable `enum` fields with `includeIfNull: false`.
+
+## 6.5.2
+
+- Better handling of `null` when encoding `enum` values or values with
+  conversions.
+
+## 6.5.1
+
+- Fixed `BigInt`, `DateTime`, and `Uri` support for `JsonKey.defaultValue` with
+  a function value.
+
+## 6.5.0
+
+- Allow constructors to be passed to `JsonKey` parameters that support
+  `Function` types.
+- Accept `Function` values for `JsonKey.defaultValue`. The provided
+  `Function` will be invoked for the default value if the target JSON element is
+  missing or `null`.
+
+## 6.4.1
+
+- Fixed a bug when an `@JsonSerializable` class uses a mixin with fields.
+- Added more documentation `@JsonEnum`.
+
 ## 6.4.0
 
 - Add support for `JsonEnum.valueField` which allows specifying a field in an